+++ workspaces/jfrog-artifactory/plugins/jfrog-artifactory/src/__fixtures__/mockTags.ts
@@ -396,6 +396,10 @@ export const mockTags = {
           },
         },
       ],
+      pageInfo: {
+        hasNextPage: false,
+        endCursor: null,
+      },
     },
   },
 };