+++ workspaces/servicenow/plugins/servicenow/src/components/Servicenow/EntityServicenowContent.tsx
@@ -234,7 +234,23 @@ export const EntityServicenowContent = () => {
         page={pageNumber}
         onPageChange={handlePageChange}
         onRowsPerPageChange={handleRowsPerPageChange}
-        labelRowsPerPage={null}
+        labelRowsPerPage={
+          <span
+            style={{
+              position: 'absolute',
+              width: '1px',
+              height: '1px',
+              padding: 0,
+              margin: '-1px',
+              overflow: 'hidden',
+              clip: 'rect(0, 0, 0, 0)',
+              whiteSpace: 'nowrap',
+              border: 0,
+            }}
+          >
+            {t('table.rowsPerPage')}
+          </span>
+        }
         showFirstButton
         showLastButton
       />