+++ workspaces/cicd-statistics/plugins/cicd-statistics/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-cicd-statistics
 
+## 0.19.0
+
+### Minor Changes
+
+- 283b19b: Backstage version bump to v1.53.0
+
 ## 0.18.0
 
 ### Minor Changes