+++ workspaces/github/plugins/github-discussions/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage-community/plugin-github-discussions
 
+## 1.3.0
+
+### Minor Changes
+
+- d130786: Backstage version bump to v1.53.0
+
+### Patch Changes
+
+- Updated dependencies [d130786]
+  - @backstage-community/plugin-github-discussions-common@0.15.0
+
 ## 1.2.0
 
 ### Minor Changes