+++ workspaces/kiali/plugins/kiali/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage-community/plugin-kiali
 
+## 1.51.1
+
+### Patch Changes
+
+- e9247c7: Bumped Backstage dependencies to 1.49.4 for Red Hat Developer Hub 1.10 compatibility.
+- Updated dependencies [e9247c7]
+  - @backstage-community/plugin-kiali-common@0.11.1
+  - @backstage-community/plugin-kiali-react@0.7.3
+
 ## 1.51.0
 
 ### Minor Changes