+++ workspaces/splunk/plugins/splunk-on-call/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-splunk-on-call
 
+## 0.25.0
+
+### Minor Changes
+
+- 33dd2c6: Backstage version bump to v1.53.0
+
 ## 0.24.0
 
 ### Minor Changes